Refactor pure utility helpers into modules
